add report

+++ exp1/exp1-1.cpp
@@ -2,20 +2,20 @@
 #include <iostream>
 using namespace std;
 
-float xn(int k) {
-    float ans = 1.0 / k;
+double xn(int k) {
+    double ans = 1.0 / k;
     ans = k % 2 == 0 ? -ans : ans;
     return ans;
 }
 
 int main() {
     int n = 1;
-    float temSum = 0.0;
-    float target = 0.693147190546;
-    float threshold = 0.5 * 1e-4;
+    double temSum = 0.0;
+    double target = 0.693147190546;
+    double threshold = 0.5 * 1e-4;
     while (true) {
         temSum += xn(n);
-        float delta = temSum - target;
+        double delta = temSum - target;
         if (fabs(delta) < threshold) {
             cout << "n = " << n << endl;
             break;
